requirements.txt
transformers==4.42.3
datasets==2.20.0
pandas==2.2.2
numpy==1.26.4
torch==2.1.2
gc-python-utils==0.0.1
tqdm==4.66.4
safetensors==0.4.3
accelerate==0.32.1
wandb==0.17.4
omegaconf==2.3.0
einops==0.8.0
sentencepiece==0.2.0
Pillow==10.3.0
torchvision==0.15.2